Get token to keep session - Google and Form
