+++ src/views/Register/Register.jsx
@@ -1,39 +1,16 @@
-import React, { useState } from 'react'
-import { useDispatch } from 'react-redux'
-import { createUser, googleAuth } from '../../store/actions/user'
+import React from 'react'
+import { createUser} from '../../store/actions/user'
 import { GoogleLogin } from '@react-oauth/google'
+import useAuth from '../../hooks/useAuth'
 
 const Register = () => {
-
-    const dispatch = useDispatch()
-
-    const [values, setValues] = useState({})
-
-    const handleChange = (e) => {
-        const { name, value } = e.target
-        setValues({ ...values, [name]: value })
-    }
-
-    const handleSubmit = (e) => {
-        e.preventDefault()
-        dispatch(createUser(values))
-            .then(res => {
-                console.log("RES", res)
-            })
-    }
-
-    const handleOnSuccessGoogle = (res) => {
-        dispatch(googleAuth(res))
-            .then(data => {
-                console.log('RES', data)
-            })
-    }
+    const { handleSubmit, handleChange, handleOnSuccessGoogle} = useAuth()
 
     return (
         <>
             <div>Register</div>
             <form
-                onSubmit={handleSubmit}
+                onSubmit={(e)=>handleSubmit(e, createUser)}
                 onChange={handleChange}
                 action=""
             >
